requirements.txt
DrissionPage
pytest
allure-pytest
PyYAML
python-dotenv
requests
loguru